extract buttons to separate component
